Added Posts for events and blog with comments and replies Added a helper to check if user is logged in Return a message when a widget is not available user_id() helper only returns default when user not found Removed ourzobia logo as default

+++ app/Helpers/currency_helper.php
@@ -15,7 +15,7 @@ if (! function_usable('money'))
 		if ($amount < 0) {
 			$amount = 0;
 		}
-		return $span_o.currency(3, $currency??my_config('site_currency')).$span_c . number_format((int)$amount, 2);
+		return $span_o.currency(3, $currency??my_config('site_currency')).$span_c . number_format((float)$amount, 2);
 	}
 }
